refactor(gulp) rename gulp tasks file
